[#217] update for dev Kinesis
IAM permissions have been granted for new dev Kinesis streams. This
updates env vars so everything works out of the box.

It seemed to be necessary to set GOOGLE_GEOCODE_API_KEY, unlike when
transforming locally.

Use new REDIS_URL env var, rather than REDIS_HOST, REDIS_PORT.

Remove error rescues in Apps::Transformer, Apps::TransformerBulk,
Clients::GoogleGeocoderClient. These were probably added for a good
reason, most likely because of errors in the datasources, but they lead
to errors being hidden and misreported, which wastes development time.
This particularly was an issue for Google Geocoding API, before the key
was set, and also when developing IAM permissions for dev Kinesis. It
might well be necessary to make a compromise to this if issues are found
in the datasources.